pyproject.toml
[tool.poetry]
name = "memc-load"
version = "0.1.0"
description = "OTUS Homework #9. Memc Load"
authors = ["inna-tuzhikova <inna-tuzhikova@github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-memcached = "^1.62"
protobuf = "^4.25.2"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
flake8 = "^7.0.0"
isort = "^5.13.2"
flake8-pyproject = "^1.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 80
multi_line_output = 3
include_trailing_comma = true
use_parentheses = true
ensure_newline_before_comments = true

[tool.flake8]
ignore = ["W503", "F811", "E722"]
max-line-length = 80
